add core tools stream
