+++ runtime/funasr_api/example.py
@@ -14,16 +14,18 @@ def recognizer_example():
         uri="wss://www.funasr.com:10096/"
     )
     text=rcg.rec_file("asr_example.mp3")
-    print("asr_example.mp3 text=",text)
+    print("asr_example.mp3 result=",text)
+ 
 
 def recognizer_stream_example():
+
+    rcg = FunasrApi(
+        uri="wss://www.funasr.com:10096/"
+    )
     #define call_back function for msg 
     def on_msg(msg):
-       print("stream_example msg=",msg)
-    rcg = FunasrApi(
-        uri="wss://www.funasr.com:10096/",msg_callback=on_msg
-    )
-    rcg.create_connection()
+       print("stream msg=",msg)
+    stream=rcg.create_stream(msg_callback=on_msg)
     
     wav_path = "asr_example.wav"
 
@@ -32,21 +34,20 @@ def recognizer_stream_example():
         
     # use FunasrApi's audio2wav to covert other audio to PCM if needed
     #import os
+    #from funasr_tools import FunasrTools
     #file_ext=os.path.splitext(wav_path)[-1].upper()
     #if not file_ext =="PCM" and not file_ext =="WAV":
-    #       audio_bytes=rcg.audio2wav(audio_bytes)
+    #       audio_bytes=FunasrTools.audio2wav(audio_bytes)
     
     stride = int(60 * 10 / 10 / 1000 * 16000 * 2)
     chunk_num = (len(audio_bytes) - 1) // stride + 1
 
     for i in range(chunk_num):
-
         beg = i * stride
         data = audio_bytes[beg : beg + stride]
-
-        rcg.feed_chunk(data)
-    msg=rcg.get_result()
-    print("asr_example.wav text=",msg)
+        stream.feed_chunk(data)
+    final_result=stream.wait_for_end()
+    print("asr_example.wav stream_result=",final_result)
     
     
 if __name__ == "__main__":
@@ -54,7 +55,7 @@ if __name__ == "__main__":
     print("example for Funasr_websocket_recognizer")
  
     recognizer_stream_example()
-
+   
     recognizer_example()
